% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emmie_notes.R
\name{emmie_notes}
\alias{emmie_notes}
\title{Notes function}
\usage{
emmie_notes(
  file_path = "C:/Users/emmie/OneDrive/Bureaublad/r_course/week_3/Notes.txt",
  keyword = NULL
)
}
\arguments{
\item{file_path}{This is the path to the file with the notes.}

\item{keyword}{This is where you put in the keyword to help filter the notes. If it is not provided it will give the whole notes file.}
}
\value{
This function returns either the whole notes file, or the lines that have the matching keyword.
}
\description{
This function searches my notes, for the search you can specify a keyword. Happy cheating!
}
