% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emmie_scatterplot.R
\name{emmie_scatterplot}
\alias{emmie_scatterplot}
\title{Scatterplot Function}
\usage{
emmie_scatterplot(dataset, x_name, y_name)
}
\arguments{
\item{dataset}{This includes the dataset to use}

\item{x_name}{This is the name of the column for the x-axis}

\item{y_name}{This is the name of the column for the y-axis}
}
\value{
This function returns a ggplot scatterplot
}
\description{
This function creates a scatterplot with Emmie's design, absolutely beautiful
}
